2024/4/24
new version

+++ src/main/java/com/cpt202/demo/repository/ManagerRepo.java
@@ -1,11 +1,14 @@
 package com.cpt202.demo.repository;
 
+import com.cpt202.demo.entity.Manager;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.cpt202.demo.entity.Manager;
+//codeby:Haoyu.li
 
 @Repository
-public interface ManagerRepo extends JpaRepository<Manager, Integer> {
-    // Custom queries (if needed)
+
+public interface ManagerRepo extends JpaRepository<Manager,Long> {
+    //Custom queries
+
 }
